requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sse-starlette>=1.6.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Config
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI Agents
smolagents[litellm]>=1.24.0
datasets>=4.8.3
markdownify>=1.2.2
duckduckgo-search>=8.1.1
spaces>=0.47.0
gradio-tools>=0.0.9
langchain>=1.2.12
langchain-community>=0.4.1
langchain-classic>=1.0.3
langchain-huggingface>=1.2.1
sentence-transformers>=3.0.0
faiss-cpu>=1.13.2
transformers<5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.20.0
